requirements.txt
addict==2.4.0
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
distlib==0.3.4
docker-pycreds==0.4.0
entrypoints==0.4
filelock==3.4.1
gitdb==4.0.9
GitPython==3.1.18
idna==3.3
imagecodecs==2020.5.30
imageio==2.15.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.3.1
matplotlib==3.3.4
nest-asyncio==1.5.5
networkx==2.5.1
numpy==1.19.5
opencv-python==4.5.5.64
pandas==1.1.5
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pipenv==2022.1.8
platformdirs==2.4.0
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
Pygments==2.12.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==23.0.0
requests==2.27.1
scikit-image==0.17.2
scipy==1.5.4
sentry-sdk==1.5.12
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
tifffile==2020.9.3
torch==1.10.1
torchvision==0.11.2
tornado==6.1
tqdm==4.64.0
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.9
virtualenv==20.13.3
virtualenv-clone==0.5.7
wandb==0.12.16
wcwidth==0.2.5
zipp==3.6.0